+++ Docs/ReleaseNotes.md
@@ -7,6 +7,9 @@ Note that the hardware framework projects have their own release cycle and relea
 
 ## vNext
 
+- New Loopback sample to test FPGA connectivity and Hastlayer Hardware Framework resource usage.
+- Various smaller bugfixes and improvements.
+
 Note that running Hastlayer now requires Visual Studio 2017 or greater (any edition will work).